feat: add navigation to login, refactor folder structure
