+++ src/components/Account/AccountBlocks/ChangePassword/ChangePassword.tsx
@@ -5,7 +5,7 @@ import s from './ChangePassword.module.scss';
 
 import Button from 'src/components/Button/Button';
 import InputPassword from 'src/components/InputPassword/InputPassword';
-import { changePasswSchema } from 'src/components/Account/AccountBlocks/ChangePassword/changePassword.schema';
+import { changePasswSchema } from 'src/schemas/changePassword.schema';
 import ModalMsg from 'src/components/ModalMsg/ModalMsg';
 
 interface InitValues {
